Initial split functions (helpers) out to separate file
